refactor NetworkFee response objects; fix Get Trades
