+++ src/main/java/trade/wayruha/whitebit/dto/FeeInfo.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -17,23 +16,6 @@ public class FeeInfo {
   private String ticker;
   private String name;
   private List<String> providers;
-  private Fee withdraw;
-  private Fee deposit;
-
-  @Data
-  public static class Fee {
-      @JsonAlias("max_amount")
-      private BigDecimal maxAmount;
-      @JsonAlias("min_amount")
-      private BigDecimal minAmount;
-      private BigDecimal fixed;
-      private FlexibleFee flex;
-
-      @Data
-      public static class FlexibleFee {
-          private BigDecimal percent;
-          private BigDecimal min_fee;
-          private BigDecimal max_fee;
-      }
-  }
+  private NetworkFee withdraw;
+  private NetworkFee deposit;
 }
